Constructor Defintions and SearchCatalog Changes
Added the Constructor Definition for both defuatl and overloaded Constructors
from both Book and Library Classes.

Everything runs as expected besides for the SearchCatalog function.
Returns that no titles are found for keywords that are indeed in the book catalog
Changed the handling of the else statement by using a boolean flag if the keyword
isn't found in the title.

Even after implementing it this way, The problem persists. May have to do with the
if statment calling .find() more than once or just being too complex.

Can't be leftover newline characters from previous user inputs in the program, as
cin.ignore() is already implemented to discard any leftover newline characters.