LoadCatalog Logic Redesign
Changed the logic for LoadCatalog without the use of substr and getline statements.
Instead, I used a string placeholder variable to manually load in each character
of te vraible until the delimiter is hit. Then the string holder gets set as
the variable, using stoi and stod where appropiate. This way, coding standards
are adhered to.